coletor/requirements.txt
python-twitter
pymongo
